Brake Bias

+++ Telemetry/Receiver/Receiver.ino
@@ -65,7 +65,8 @@ void setup() {
  
 void loop() {
   // CSV format Serial Print
-  brakeBias = (.99 * (double)telemetry.BrakeFront) / ((.99 * (double)telemetry.BrakeFront) + (0.79 * (double)telemetry.BrakeRear));
+  
+  brakeBias = (.99 * telemetry.BrakeFront) / ((.99 * telemetry.BrakeFront) + (0.79 * telemetry.BrakeRear));
   Serial.printf("%f, %f, %f, %f, %f, %f, %f, ", telemetry.RPM, telemetry.TPS, telemetry.FOT, telemetry.IA, telemetry.Lam, telemetry.AirT, telemetry.CoolT);
   Serial.printf("%f, %f, %f, %f, %f, %f, %f, %f, ", telemetry.Lat, telemetry.Lng, telemetry.Speed, telemetry.OilP, telemetry.FuelP, telemetry.FLTemp, telemetry.FRTemp, telemetry.RLTemp);
   Serial.printf("%f, %f, %f, %f, %f, %f, %f, %d, ", telemetry.RRTemp, telemetry.FRPot, telemetry.FLPot, telemetry.RRPot, telemetry.RLPot, telemetry.BrakeFront, telemetry.BrakeRear, brakeBias);
